feat: Working cube localiser, renamed yolo_3d_to_marker to cube_localiser. TF fully working, cube avg position now accounted for. Only depends on bounding boxes information from model now. Added more coloured boxes to auto_cubes.sdf
