+++ src/ros/rover/nova_gazebo/worlds/auto_cubes.sdf
@@ -64,6 +64,45 @@
       </ode>
     </physics>
 
+    <model name='red_cube'>
+      <pose>6.5 -6.0 2.5 0 0 0</pose>
+      <link name='link'>
+        <visual name='visual'>
+          <geometry>
+            <box>
+              <size>0.2 0.2 0.2</size>
+            </box>
+          </geometry>
+            <material>
+              <ambient>0.5 0.1 0.1 1</ambient>   <!-- Dark red ambient -->
+              <diffuse>1 0 0 1</diffuse>         <!-- Pure red under light -->
+              <specular>0.2 0.2 0.2 1</specular> <!-- Slight shine -->
+              <emissive>0 0 0 1</emissive>       <!-- No self-glow -->
+            </material>
+        </visual>
+        <collision name='collision'>
+          <geometry>
+            <box>
+              <size>0.2 0.2 0.2</size>
+            </box>
+          </geometry>
+          <max_contacts>10</max_contacts>
+          <surface>
+            <contact>
+              <ode/>
+            </contact>
+            <bounce/>
+            <friction>
+              <torsional>
+                <ode/>
+              </torsional>
+              <ode/>
+            </friction>
+          </surface>
+        </collision>
+      </link>
+    </model>
+
     <model name='blue_cube'>
       <pose>5.5 -4.0 0.5 0 0 0</pose>
       <link name='link'>
@@ -102,6 +141,124 @@
         </collision>
       </link>
     </model>
+
+    <model name='green_cube'>
+      <pose>7.5 -2.0 3.5 0 0 0</pose>
+      <link name='link'>
+        <visual name='visual'>
+          <geometry>
+            <box>
+              <size>0.2 0.2 0.2</size>
+            </box>
+          </geometry>
+            <material>
+                <ambient>0.1 0.5 0.1 1</ambient>   <!-- Dark green ambient -->
+                <diffuse>0 1 0 1</diffuse>         <!-- Pure green under light -->
+                <specular>0.2 0.2 0.2 1</specular> <!-- Slight shine -->
+                <emissive>0 0 0 1</emissive>       <!-- No self-glow -->
+            </material>
+        </visual>
+        <collision name='collision'>
+          <geometry>
+            <box>
+              <size>0.2 0.2 0.2</size>
+            </box>
+          </geometry>
+          <max_contacts>10</max_contacts>
+          <surface>
+            <contact>
+              <ode/>
+            </contact>
+            <bounce/>
+            <friction>
+              <torsional>
+                <ode/>
+              </torsional>
+              <ode/>
+            </friction>
+          </surface>
+        </collision>
+      </link>
+    </model>
+
+    <model name='white_cube'>
+      <pose>3.5 -4.0 2.5 0 0 0</pose>
+      <link name='link'>
+        <visual name='visual'>
+          <geometry>
+            <box>
+              <size>0.2 0.2 0.2</size>
+            </box>
+          </geometry>
+            <material>
+                <ambient>0.6 0.6 0.6 1</ambient>   <!-- Light gray ambient -->
+                <diffuse>1 1 1 1</diffuse>         <!-- Pure white under light -->
+                <specular>0.3 0.3 0.3 1</specular> <!-- Slight shine -->
+                <emissive>0 0 0 1</emissive>       <!-- No self-glow -->
+            </material>
+        </visual>
+        <collision name='collision'>
+          <geometry>
+            <box>
+              <size>0.2 0.2 0.2</size>
+            </box>
+          </geometry>
+          <max_contacts>10</max_contacts>
+          <surface>
+            <contact>
+              <ode/>
+            </contact>
+            <bounce/>
+            <friction>
+              <torsional>
+                <ode/>
+              </torsional>
+              <ode/>
+            </friction>
+          </surface>
+        </collision>
+      </link>
+    </model>
+    <!--
+    <model name='black_cube'>
+      <pose>8.5 -2.0 5.5 0 0 0</pose>
+      <link name='link'>
+        <visual name='visual'>
+          <geometry>
+            <box>
+              <size>0.2 0.2 0.2</size>
+            </box>
+          </geometry>
+            <material>
+              <ambient>0.05 0.05 0.05 1</ambient>
+              <diffuse>0 0 0 1</diffuse>         
+              <specular>0 0 0 1</specular>      
+              <emissive>0 0 0 1</emissive>       
+            </material>
+        </visual>
+        <collision name='collision'>
+          <geometry>
+            <box>
+              <size>0.2 0.2 0.2</size>
+            </box>
+          </geometry>
+          <max_contacts>10</max_contacts>
+          <surface>
+            <contact>
+              <ode/>
+            </contact>
+            <bounce/>
+            <friction>
+              <torsional>
+                <ode/>
+              </torsional>
+              <ode/>
+            </friction>
+          </surface>
+        </collision>
+      </link>
+    </model>
+    -->
 
   </world>
 </sdf>
