+++ src/ros/rover/nova_utils/nova_utils/cube_localiser.py
@@ -8,9 +8,10 @@ as a transform for cube localisation
 NODE: yolo_3d_to_marker
 TOPICS:
   - subscriber: /yolo/detections [yolo_msgs/msg/DetectionArray]
-  - subscriber: /oak/depth/points [sensor_msgs/msg/PointCloud2]
-  - publisher: /yolo/cubes [visualization_msgs.MarkerArray]
-  - publisher: /tf
+  - subscriber: /oak/depth       [sensor_msgs/msg/Image]
+  - subscriber: /oak/camera_info [sensor_msgs/msg/CameraInfo]
+  - publisher: /yolo/cubes       [visualization_msgs.MarkerArray]
+  - publisher: /tf               [geometry_msgs/msg/TransformStamped]
 SERVICES: None
 ACTIONS: None
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -20,7 +21,9 @@ CREATION:	20/02/2025
 EDITED:		20/02/2025
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 TODO:
- - Refine max_std_dev param
+ - Refine the statistical analysis
+ - Refine the standard dev and min samples params
+ - Convert to work for oak camera
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 """
 import rclpy
@@ -29,9 +32,10 @@ from rclpy.duration import Duration
 from rclpy.qos import QoSHistoryPolicy, QoSDurabilityPolicy, QoSProfile
 
 from visualization_msgs.msg import MarkerArray, Marker
-from geometry_msgs.msg import Pose, Point, Vector3, TransformStamped
+from geometry_msgs.msg import Pose, Vector3, TransformStamped
 from yolo_msgs.msg import DetectionArray, Detection, BoundingBox3D
 from sensor_msgs.msg import Image, CameraInfo
+from builtin_interfaces.msg import Time
 
 from tf2_ros.transform_broadcaster import TransformBroadcaster
 from tf2_ros import Buffer, TransformListener
@@ -44,37 +48,50 @@ import numpy as np
 
 from typing import Dict, List, Tuple, TypeVar
 T = TypeVar('T')
+type Point = Tuple[float, float, float] # point = (x,y,z)
+
 
 COLORS = {'red':[1.0,0.0,0.0], 'green':[0.0,1.0,0.0], 'blue':[0.0,0.0,1.0], 'white':[1.0,1.0,1.0]}
 DEFAULT_QUATERNION = [0.0, 0.0, 0.0, 1.0]
-RVIZ_CUBE_SIZE = 0.15
 
-DETECTION_TOPIC = "/detections"
-MARKER_TOPIC = "/cubes"
-DETECTION_TYPES = ["detected"]
-TARGET_FRAME = "map"
+# change these topics using remapping in launch file
+DEPTH_IMAGE_TOPIC = "/oak/depth"
+DEPTH_INFO_TOPIC = "/oak/camera_info"
+DETECTION_TOPIC = "/yolo/detections"
+MARKER_TOPIC = "/yolo/cubes"
+
 
 class DetectionTransformer(Node):
     def __init__(self):
         super().__init__("detection_transformer")
-        namespace = self.get_namespace() 
-        namespace = "/yolo"
 
-        self.declare_parameter('target_frame', 'map')
-        self.depth_image_units_divisor = 1
-        self.maximum_detection_threshold = 0.3
+        # variables affecting detection of cubes
+        self.depth_image_units_divisor = self.declare_parameter('depth_image_units_divisor', 1.0).get_parameter_value().double_value
+        self.maximum_detection_threshold = self.declare_parameter('maximum_detection_threshold', 0.3).get_parameter_value().double_value
+        self.significant_threshold = self.declare_parameter('significant_threshold', 0.01).get_parameter_value().double_value
+
+        # variables affecting cube markers in rviz
+        self.use_markers = self.declare_parameter('use_markers', True).get_parameter_value().bool_value
+        self.marker_ns = self.declare_parameter('marker_ns', 'detected_cubes').get_parameter_value().string_value
+        self.marker_duration = self.declare_parameter('marker_duration', 1.0).get_parameter_value().double_value
+        self.marker_size = self.declare_parameter('marker_size', 0.15).get_parameter_value().double_value
+
+        # variables to determine frame of map and camera
+        self.map_frame = self.declare_parameter('map_frame', 'map').get_parameter_value().string_value
+        self.camera_frame = self.declare_parameter('camera_frame', 'camera_link').get_parameter_value().string_value
+
+        # timer period to publish cube transforms in seconds
+        self.tf_publisher_timer_period = self.declare_parameter('tf_publisher_timer_period', 0.1).get_parameter_value().double_value
+
+        # variables for statistical analysis
+        self.min_samples = self.declare_parameter('min_samples', 5).get_parameter_value().integer_value
+        self.max_std_dev = self.declare_parameter('max_std_dev', 0.2).get_parameter_value().double_value
 
         self.transform_broadcaster = TransformBroadcaster(self)
         self.tf_buffer = Buffer()
         self.tf_listener = TransformListener(self.tf_buffer, self)
 
         self.cv_bridge = CvBridge()
-
-        self.publisher = self.create_publisher(
-            MarkerArray,
-            namespace + MARKER_TOPIC,
-            10
-        )
 
         self.default_qos_profile = QoSProfile(
             reliability=1,
@@ -83,37 +100,77 @@ class DetectionTransformer(Node):
             depth=1,
         )
 
+        # subscribe to the topics
         self.depth_sub = message_filters.Subscriber(
-            self, Image, "/oak/depth", qos_profile=self.default_qos_profile
+            self, Image, DEPTH_IMAGE_TOPIC, qos_profile=self.default_qos_profile
         )
         self.depth_info_sub = message_filters.Subscriber(
-            self, CameraInfo, "/oak/camera_info", qos_profile=self.default_qos_profile
+            self, CameraInfo, DEPTH_INFO_TOPIC, qos_profile=self.default_qos_profile
         )
         self.detections_sub = message_filters.Subscriber(
-            self, DetectionArray, "/yolo/detections"
+            self, DetectionArray, DETECTION_TOPIC
         )
-
+        # synchronise information from topics and run function upon all information received
         self._synchronizer = message_filters.ApproximateTimeSynchronizer(
-            (self.depth_sub, self.depth_info_sub, self.detections_sub), 10, 0.5
-        )
+            (self.depth_sub, self.depth_info_sub, self.detections_sub), 10, 0.5)
         self._synchronizer.registerCallback(self.on_detections)
 
+        # publish to the marker topic
+        self.publisher = self.create_publisher(MarkerArray, MARKER_TOPIC, 10)
+
+        self.detected_cubes : Dict[str, List[Point]] \
+            = {'red':[], 'green':[], 'blue':[], 'white':[]}
+        
+        # run the callback function every timer_period
+        self.create_timer(self.tf_publisher_timer_period, self.publish_cubes)
+
+
     def on_detections(self, depth_msg: Image, depth_info_msg: CameraInfo, detections_msg: DetectionArray) -> None:
+        """Process and publish detected cubes"""
         detections = self.process_detections(depth_msg, depth_info_msg, detections_msg)
 
-        msg = MarkerArray()
-
-        for i, detection in enumerate(detections):
-            marker = self.get_marker(i, detection[0], detection[1], detections_msg.header.stamp, "camera_link")
-            msg.markers.append(marker)
+        # Markers
+        if (self.use_markers):
+            msg = MarkerArray()
+            detection: Tuple[str, Point]
+            for i, detection in enumerate(detections):
+                marker = self.get_marker(i, detection[0], detection[1], detections_msg.header.stamp, self.map_frame)
+                msg.markers.append(marker)
+            
+            self.publisher.publish(msg)
         
-        self.publisher.publish(msg)
+        # Transforms
+        detection: Tuple[str, Point]
+        for i, detection in enumerate(detections):
+            self.detected_cubes[detection[0]].append(detection[1])
 
+    def publish_cubes(self) -> None:
+        """Publish the current transform of all detected cubes"""
+        for color, points in self.detected_cubes.items():
+            if len(points) >= self.min_samples:
+                clean_points = self.remove_outlier_pos(points)  # remove outliers from all the cube points
+
+                if len(clean_points) >= self.min_samples:       # ensure there is enough samples
+                    self.get_logger().debug(f"Validating consistency of target {color}")
+                    avg_pos = np.mean(clean_points, axis=0)   # Calculate the average position of the block
+                    std_dev = np.std(clean_points, axis=0)    # Calculate the standard deviation of the block's position
+
+                    if np.all(std_dev < self.max_std_dev):      # Check that the standard deviation is small enough to be considered a confirmed block
+                        self.get_logger().debug(f"Confirmed target {color} consistent pos at {avg_pos}")
+                        self.publish_tf(color, avg_pos, self.get_clock().now().to_msg())
+
+                    else:
+                        self.get_logger().debug(f"Target {color} is not consistent enough.")
+                else:
+                    self.get_logger().debug(f"{color} has not enough samples to confirm")
+                
 
     def process_detections(self, depth_msg: Image, depth_info_msg: CameraInfo, detections_msg: DetectionArray) -> List[Tuple[str,Point]]:
-        # check if there are detections
+        """Process detections into a list of points with their respective colour"""
         if not detections_msg.detections:
             return []
+
+        self.get_logger().debug(f"Processing new detections")
 
         new_detections = []
         depth_image = self.cv_bridge.imgmsg_to_cv2(depth_msg, desired_encoding="32FC1")
@@ -121,14 +178,17 @@ class DetectionTransformer(Node):
         for detection in detections_msg.detections:
             position = self.convert_bb_to_point(depth_image, depth_info_msg, detection)
             if position is not None:
-                # transform point from image frame to map frame 
-                new_detections.append((detection.class_name, position))
+                new_position = self.tf_to_map(position, detections_msg.header.stamp)
+                if new_position is not None:
+                    new_detections.append((detection.class_name, new_position))
 
         return new_detections
 
     def convert_bb_to_point(self, depth_image: np.ndarray, depth_info: CameraInfo, detection: Detection) -> Point | None:
-        """Converts the bounding box center to a point relative to the image frame"""
-
+        """ Converts the bounding box center to a point relative to the image frame
+            Modified from convert_bb_to_3d in https://github.com/mgonzs13/yolo_ros/blob/main/yolo_ros/yolo_ros/detect_3d_node.py
+        """
+        self.get_logger().debug(f"Calculating cube world position relative to image frame")
         center_x = int(detection.bbox.center.position.x)
         center_y = int(detection.bbox.center.position.y)
         size_x = int(detection.bbox.size.x)
@@ -175,29 +235,78 @@ class DetectionTransformer(Node):
         # rotate 90 deg around x and -90 deg around z from camera_link
         return (z, -1*x, -1*y)
 
-    def get_marker(self, id:int, color:str, point: Tuple[float, float, float], stamp, frame:str) -> Marker:
+    def quaternion_to_rotation_matrix(self, q: Tuple[float, float, float, float]) -> np.array:
+        """Convert a quaternion (x, y, z, w) into a 3x3 rotation matrix."""
+        x, y, z, w = q
+        return np.array([
+            [1 - 2 * (y ** 2 + z ** 2), 2 * (x * y - z * w), 2 * (x * z + y * w)],
+            [2 * (x * y + z * w), 1 - 2 * (x ** 2 + z ** 2), 2 * (y * z - x * w)],
+            [2 * (x * z - y * w), 2 * (y * z + x * w), 1 - 2 * (x ** 2 + y ** 2)]
+        ])
+
+    def tf_to_map(self, camera_to_cube: Point, stamp: Time) -> Point | None:
+        """Calculates the position tf from map to cube"""
+        self.get_logger().debug(f"Calculating map to cube tf")
+        try:
+            map_to_camera = self.tf_buffer.lookup_transform(self.map_frame, self.camera_frame, stamp).transform
+
+            # rotate the map_to_camera position by its orientation 
+            rot_matrix = self.quaternion_to_rotation_matrix([map_to_camera.rotation.x,map_to_camera.rotation.y, map_to_camera.rotation.z, map_to_camera.rotation.w])
+            rotated_translation = np.dot(rot_matrix, camera_to_cube)
+
+            # apply the camera_to_cube tf to the rotated map_to_camera
+            return (map_to_camera.translation.x+rotated_translation[0], map_to_camera.translation.y+rotated_translation[1], map_to_camera.translation.z+rotated_translation[2])
+        except Exception as e:
+            self.get_logger().debug(f"Error in calculating map to cube tf {e}")
+            return None
+
+    def get_marker(self, id:int, color:str, point: Point, stamp: Time, frame:str) -> Marker:
         """Returns a marker derived from the detection"""
         marker = Marker()
         marker.pose.position.x, marker.pose.position.y, marker.pose.position.z = point
         marker.pose.orientation.x, marker.pose.orientation.y, marker.pose.orientation.z, marker.pose.orientation.w = DEFAULT_QUATERNION
 
         marker.type = Marker.CUBE
-        marker.scale.x = RVIZ_CUBE_SIZE
-        marker.scale.y = RVIZ_CUBE_SIZE
-        marker.scale.z = RVIZ_CUBE_SIZE
+        marker.scale.x = self.marker_size
+        marker.scale.y = self.marker_size
+        marker.scale.z = self.marker_size
         marker.color.r = COLORS[color][0]
         marker.color.g = COLORS[color][1]
         marker.color.b = COLORS[color][2]
         marker.color.a = 1.0
 
-        marker.lifetime = Duration(seconds=0.3).to_msg()
-        marker.ns = "detected"
+        marker.lifetime = Duration(seconds=self.marker_duration).to_msg()
+        marker.ns = self.marker_ns
         marker.id = id
 
         marker.header.stamp = stamp
         marker.header.frame_id = frame
 
         return marker
+
+    def remove_outlier_pos(self, pos_vals: List[T]) -> List[T]:
+        """
+        Removes any outlier positions from the list of positions. An outlier is defined as a position that is
+        more than 3 standard deviations away from the mean.
+        """
+        mean = np.mean(pos_vals, axis=0)
+        std_dev = np.std(pos_vals, axis=0)
+
+        return [pos for pos in pos_vals if np.all(np.abs(pos - mean) < 2 * std_dev)]
+
+    def publish_tf(self, color:str, position: Point, stamp: Time) -> None:
+        """Publish the transform of a confirmed cube"""
+        tfs = TransformStamped()
+        tfs.header.stamp = stamp
+        tfs.header.frame_id = self.map_frame
+        tfs.child_frame_id = color + "_cube"
+        tfs.transform.translation.x,  tfs.transform.translation.y, tfs.transform.translation.z, = position
+        tfs.transform.rotation.x, tfs.transform.rotation.y, tfs.transform.rotation.z, tfs.transform.rotation.w = DEFAULT_QUATERNION
+
+        self.transform_broadcaster.sendTransform(tfs)
+
+
+
